pyproject.toml
[project]
name = "Agent-Avatar"
version = "1.1.4"
description = "Talk to any LLM with hands-free voice interaction, voice interruption, and Live2D taking face running locally across platforms"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "chardet>=5.2.0",
    "fastapi[standard]>=0.115.8",
    "faster-whisper>=1.1.1",
    "httpx>=0.28.1",
    "langdetect>=1.0.9",
    "loguru>=0.7.2",
    "numpy>=1.26.4,<2",
    "onnxruntime>=1.20.1",
    "openai>=1.57.4",
    "pydub>=0.25.1",
    "pysbd>=0.3.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruamel-yaml>=0.18.10",
    "scipy>=1.14.1",
    "silero-vad>=5.1.2",
    "soundfile>=0.12.1",
    "tomli>=2.2.1",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "uvicorn[standard]>=0.33.0",
    "websocket-client>=1.8.0",
    "aiohttp>=3.8.0",   # Required by agent_zero_client
    "twitchio>=2.9.1",  # Twitch IRC integration
]

